src/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.9
aiofiles==23.2.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.11.0
PyMuPDF==1.22.3
tenacity==8.2.3
cryptography==42.0.2
python-jose==3.3.0
passlib==1.7.4
